requirements.txt
Pillow==5.0.0
PyTweening==1.0.3